# Use slice pooling to populate the query stream response